Move label smoothing test case
